% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhspd.R
\name{nhspd_as_sfc}
\alias{nhspd_as_sfc}
\title{Convert NHSPD table to SF}
\usage{
nhspd_as_sfc(nhspd)
}
\arguments{
\item{nhspd}{NHS Postcode Database object (loaded by \code{nhspd_load()})}
}
\description{
Converts NHSPD to a SF object by turning the OS easting and northing values to a st_point
}
