% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_ods.R
\docType{data}
\name{ODS_API_ENDPOINT}
\alias{ODS_API_ENDPOINT}
\title{ODS API Endpoint}
\format{
An object of class \code{character} of length 1.
}
\usage{
ODS_API_ENDPOINT
}
\description{
The API endpoint for the NHS Organisation Data Services: all calls to get
data should start from this URI.
}
\keyword{datasets}
