% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_ods.R
\docType{data}
\name{ODS_API_RESULTS_LIMIT}
\alias{ODS_API_RESULTS_LIMIT}
\title{ODS API Results Limit}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ODS_API_RESULTS_LIMIT
}
\description{
This is the maximum number of results we should try to load from the API.
Note that this is limited to 1000
}
\keyword{datasets}
