% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_get_roles.R
\name{ods_get_roles}
\alias{ods_get_roles}
\title{ODS Roles}
\usage{
ods_get_roles()
}
\value{
A tibble containing the roles that can be used to query the ODS
organisations api (with \code{\link{ods_get_organisations}})
}
\description{
This function queries the \href{https://digital.nhs.uk/services/organisation-data-service/guidance-for-developers/roles-endpoint}{ODS Roles API}.
}
\details{
Each organisation that is in the ODS organisations API will have a primary
role (e.g. NHS TRUST) and can have secondary roles. These role ids can be
used to query the Organisations api.
}
\examples{
\dontrun{
ods_get_roles()
}
}
