% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_api_results_limit.R
\name{ods_api_results_limit}
\alias{ods_api_results_limit}
\title{ODS API Results Limit}
\usage{
ods_api_results_limit(limit)
}
\arguments{
\item{limit}{an integer between 1 and 1000}
}
\value{
the currently set limit, or the updated limit
}
\description{
Get's, or set's, the current limit of results to be returned by the ODS API.
If a value is provided for the limit argument then the value is updated. If a
value is not provided then the current limit value is returned.
}
\examples{
# get the current limit
ods_api_results_limit() # defaults to 1000

# we can update this limit to a value between 1 and 1000
ods_api_results_limit(50) # should return 50
ods_api_results_limit() # should now be 50
}
